requirements.txt
# Core dependencies
aiohttp==3.9.1
asyncio-throttle==1.0.2
python-dotenv==1.0.0
pydantic==2.5.0

# Data analysis and technical indicators
pandas==2.1.4
numpy==1.24.3
ta-lib==0.4.26
scikit-learn==1.3.2

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.13.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
colorlog==6.8.0

# Notifications
requests==2.31.0
telegram-bot-api==7.0.1
discord-webhook==1.3.1

# Web interface (optional)
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Time and date handling
python-dateutil==2.8.2
pytz==2023.3

# Configuration and validation
marshmallow==3.20.1
click==8.1.7

# Cryptography for API security
cryptography==41.0.8
websockets==12.0
